backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "storydiff"
version = "0.1.0"
description = "StoryDiff backend — data plane (Postgres + Qdrant)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0,<3",
    "alembic>=1.13,<2",
    "psycopg[binary]>=3.1,<4",
    "qdrant-client>=1.7,<2",
    "python-dotenv>=1.0,<2",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "boto3>=1.35,<2",
    "langgraph>=0.2,<1",
    "langchain-core>=0.3,<1",
    "openai>=1.0,<2",
    "httpx>=0.28,<1",
    "langgraph-checkpoint-postgres>=2.0,<3",
]

[project.optional-dependencies]
# Optional: Hugging Face `all-MiniLM-L6-v2` (requires PyTorch wheels for your platform).
embeddings-st = ["sentence-transformers>=3.0,<4"]

[dependency-groups]
dev = ["pytest>=8", "httpx>=0.28,<1", "moto[sqs]>=5.0,<6"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
